refactor: Remove interface DetectorEventPublisher and rename implementation class.
