+++ src/main/java/com/synopsys/integration/detect/tool/detector/DetectorEventPublisher.java
@@ -10,28 +10,56 @@ package com.synopsys.integration.detect.tool.detector;
 import java.io.File;
 import java.util.Set;
 
+import com.synopsys.integration.detect.workflow.event.Event;
+import com.synopsys.integration.detect.workflow.event.EventSystem;
 import com.synopsys.integration.detect.workflow.status.UnrecognizedPaths;
 import com.synopsys.integration.detector.base.DetectorEvaluationTree;
 import com.synopsys.integration.detector.base.DetectorType;
 
-public interface DetectorEventPublisher {
-    void publishApplicableCompleted(Set<DetectorType> applicableDetectors);
+public class DetectorEventPublisher {
+    private final EventSystem eventSystem;
 
-    void publishCustomerFileOfInterest(File file);
+    public DetectorEventPublisher(EventSystem eventSystem) {
+        this.eventSystem = eventSystem;
+    }
 
-    void publishDetectorsComplete(DetectorToolResult detectorToolResult);
+    public void publishApplicableCompleted(Set<DetectorType> applicableDetectors) {
+        eventSystem.publishEvent(Event.ApplicableCompleted, applicableDetectors);
+    }
 
-    void publishDiscoveryCount(Integer discoveryCount);
+    public void publishCustomerFileOfInterest(File file) {
+        eventSystem.publishEvent(Event.CustomerFileOfInterest, file);
+    }
 
-    void publishDiscoveriesCompleted(DetectorEvaluationTree detectorEvaluationTree);
+    public void publishDetectorsComplete(DetectorToolResult detectorToolResult) {
+        eventSystem.publishEvent(Event.DetectorsComplete, detectorToolResult);
+    }
 
-    void publishExtractionCount(Integer extractionCount);
+    public void publishDiscoveryCount(Integer discoveryCount) {
+        eventSystem.publishEvent(Event.DiscoveryCount, discoveryCount);
+    }
 
-    void publishExtractionsCompleted(DetectorEvaluationTree detectorEvaluationTree);
+    public void publishDiscoveriesCompleted(DetectorEvaluationTree detectorEvaluationTree) {
+        eventSystem.publishEvent(Event.DiscoveriesCompleted, detectorEvaluationTree);
+    }
 
-    void publishPreparationsCompleted(DetectorEvaluationTree detectorEvaluationTree);
+    public void publishExtractionCount(Integer extractionCount) {
+        eventSystem.publishEvent(Event.ExtractionCount, extractionCount);
+    }
 
-    void publishSearchCompleted(DetectorEvaluationTree detectorEvaluationTree);
+    public void publishExtractionsCompleted(DetectorEvaluationTree detectorEvaluationTree) {
+        eventSystem.publishEvent(Event.ExtractionsCompleted, detectorEvaluationTree);
+    }
 
-    void publishUnrecognizedPaths(UnrecognizedPaths unrecognizedPaths);
+    public void publishPreparationsCompleted(DetectorEvaluationTree detectorEvaluationTree) {
+        eventSystem.publishEvent(Event.PreparationsCompleted, detectorEvaluationTree);
+    }
+
+    public void publishSearchCompleted(DetectorEvaluationTree detectorEvaluationTree) {
+        eventSystem.publishEvent(Event.SearchCompleted, detectorEvaluationTree);
+    }
+
+    public void publishUnrecognizedPaths(UnrecognizedPaths unrecognizedPaths) {
+        eventSystem.publishEvent(Event.UnrecognizedPaths, unrecognizedPaths);
+    }
 }
